refactor(ai-service): Step 3 — split app/models/ into 4 destination bands
The word "model" was overloaded in app/models/: HTTP DTOs, ML algorithm
registry, and domain types all lived together. Each file now lands in
the band it actually belongs to (Page 6 of the diagram).

Moves:
  app/models/telemetry_models.py   →  app/domain/telemetry.py            (git mv)
  app/models/ml_algorithms.py      →  app/ml/online/algorithm_registry.py (git mv)
  app/models/segment_models.py     →  split:
       AnnotatedSegment, PredictedSegment, SegmentFeatureCatalog
         →  app/domain/segment.py
       LABEL_MAPPING, LABEL_NAME_TO_ID, MAIN_LABEL_GUIDELINES,
       LABEL_IMAGE_MAP, LABEL_CATEGORIES
         →  app/domain/labels.py
  app/models/api_models.py         →  split:
       DatasetInfo, QueryRequest, AnalysisResult, BackendCallRequest,
       HealthResponse  (HTTP boundary DTOs)
         →  app/api/schemas/query.py
       AiModelDto, ActiveModelData  (wire formats for the Node backend)
         →  app/integrations/backend/schemas.py

Net: 4 source files (76 KB) split across 6 destinations. app/models/
deleted entirely — nothing called "model" survives at the top level.

Import-graph rewrite (25 statements across 21 files): handled by a
bulk regex pass plus a few manual touch-ups for multi-line imports
and the reload pattern in ui/segment_tabs/shared.py. The archived
scikit_imitation file was updated too (`from ...models.*` → absolute
`from app.domain.telemetry`).

Validated:
  • all new symbols import cleanly (`from app.domain.labels import LABEL_MAPPING`, …)
  • SegmentFeatureCatalog still produces 88 features (touches all 3 catalogs)
  • `import app.models` raises ImportError (expected — gone)
  • app.main imports cleanly inside the container
  • /health returns 200 after uvicorn --reload picked up the change
