+++ acla_ai_service/app/api/health.py
@@ -4,7 +4,7 @@ Health check endpoints
 
 from fastapi import APIRouter
 from datetime import datetime
-from app.models import HealthResponse
+from app.api.schemas.query import HealthResponse
 from app.services.backend_service import backend_service
 from app.core import settings
