--- acla_ai_service/app/models/__init__.py
@@ -1,26 +1,0 @@
-"""
-Models module initialization
-"""
-
-from .api_models import (
-    DatasetInfo,
-    QueryRequest,
-    AnalysisResult,
-    BackendCallRequest,
-    HealthResponse,
-    AiModelDto,
-    ActiveModelData
-)
-from .telemetry_models import TelemetryFeatures, FeatureProcessor
-
-__all__ = [
-    "DatasetInfo",
-    "QueryRequest",
-    "AnalysisResult",
-    "BackendCallRequest",
-    "HealthResponse",
-    "TelemetryFeatures",
-    "FeatureProcessor",
-    "AiModelDto",
-    "ActiveModelData"
-]
