+++ acla_ai_service/app/services/llm/agent/sub_agents/label_verifier.py
@@ -19,7 +19,7 @@ from __future__ import annotations
 import logging
 from typing import Any, Dict, List, Tuple
 
-from app.models.segment_models import LABEL_MAPPING
+from app.domain.labels import LABEL_MAPPING
 from app.skill_manager import embed
 from app.services.llm.label_catalog import find_labels, get_label
 from app.services.llm.agent.framework import Agent, AgentState
