+++ acla_ai_service/app/services/llm/annotation/flows/lap.py
@@ -17,7 +17,7 @@ from typing import Any, Callable, Dict, List, Optional, Tuple
 
 import json as _json
 
-from app.models.segment_models import LABEL_MAPPING
+from app.domain.labels import LABEL_MAPPING
 from app.services.llm.label_catalog import find_labels, get_label
 from app.services.llm.skill_prompts import lap_annotation_prompt
 from app.services.llm.agent import (
