+++ acla_ai_service/app/services/segment_classifier_service.py
@@ -25,8 +25,8 @@ import copy
 from collections import defaultdict
 
 from .zarr_telemetry_store import get_shared_zarr_store
-from app.models.segment_models import AnnotatedSegment, LABEL_MAPPING, PredictedSegment, SegmentFeatureCatalog
-
+from app.domain.labels import LABEL_MAPPING
+from app.domain.segment import AnnotatedSegment, PredictedSegment, SegmentFeatureCatalog
 def compute_derived_features(df: pd.DataFrame) -> pd.DataFrame:
     """
     Computes derived features for telemetry data.
